made a few changes to transfer and faucet
